Added  oauth and authorization

+++ src/controllers/product.controller.js
@@ -1,9 +1,9 @@
-const express = require('express');
+const express = require("express");
 const router = express.Router();
 
-const Product = require('../models/product.model');
-const autheticate = require('../middlewares/authenticate');
-
+const Product = require("../models/product.model");
+const autheticate = require("../middlewares/authenticate");
+const authorize = require("../middlewares/authorize");
 router.get("", async (req, res) => {
   try {
     const product = await Product.find();
@@ -13,35 +13,48 @@ router.get("", async (req, res) => {
   }
 });
 
-router.post('',autheticate, async(req,res) => {
-    req.body.seller = req.user._id;
-    try {
-        const product = await Product.create(req.body);
-        return res.status(200).send(product);
-    } catch (error) {
-        return res.status(400).send(error);
-    }
-});
-
-router.patch("/:id",autheticate,async(req,res) =>{
-    try {
-
-        const product = await Product.findByIdAndUpdate(req.params.id,req.body,{new:true}).lean().exec();
-        console.log(req.body,product);
-        return res.status(200).send(product);
-
-    } catch (error) {
-        return res.status(400).send(error);
-        
-    }
-})
-router.delete("/:id", autheticate, async (req, res) => {
+router.post("", autheticate, async (req, res) => {
+  req.body.seller = req.user._id;
   try {
-    const product = await Product.findByIdAndDelete(req.params.id);
+    const product = await Product.create(req.body);
     return res.status(200).send(product);
   } catch (error) {
     return res.status(400).send(error);
   }
 });
+
+router.patch(
+  "/:id",
+  autheticate,
+  authorize(["admin", "seller"]),
+  async (req, res) => {
+    try {
+      const product = await Product.findByIdAndUpdate(req.params.id, req.body, {
+        new: true,
+      })
+        .lean()
+        .exec();
+        if(product){
+            console.log(req.body, product);
+        }
+      return res.status(200).send(product);
+    } catch (error) {
+      return res.status(400).send(error);
+    }
+  }
+);
+router.delete(
+  "/:id",
+  autheticate,
+  authorize(["admin", "seller"]),
+  async (req, res) => {
+    try {
+      const product = await Product.findByIdAndDelete(req.params.id);
+      return res.status(200).send(product);
+    } catch (error) {
+      return res.status(400).send(error);
+    }
+  }
+);
 
 module.exports = router;
